feat: include digital life tree graph in PDF export
Der Lebensbaum-Graph selbst fehlte bisher komplett im PDF-Export - nur
Takeaways und analoge Notizen/Fotos wurden berücksichtigt. Grund: der
Graph wurde bisher nur live per Screenshot aus dem sichtbaren
LifeTreeResult-Screen erzeugt, den es beim Export gar nicht gibt.

Die reine Graph-Darstellung (Buchheim-Walker-Layout + Knoten-Karten)
ist jetzt in LifeTreeGraphWidget ausgelagert (life_tree_result.dart
nutzt es unverändert innerhalb seines InteractiveViewer weiter). Für
den Export wird derselbe Graph zusätzlich off-screen direkt aus den
Knotendaten gerendert (ohne Constraints, analog zur bisherigen Live-
Darstellung innerhalb InteractiveViewer(constrained: false)) und
anschließend durch die bestehende Branding-Wrapper-Funktion geschickt.
